Update 05-xpath.Rmd
At the end of the document I provide an alternative solution. You specify that only the subtitle should be counted. However, I noticed that when one collects all p nodes for some reason it also collects the nodes defining the title of the article and the journalist. To circumvent that I save all nodes and the look for the children (the collection of all tags under p). I then specify that I only want the p tags. This gives me all the subtitles. I tried removing all "\n", but that did not work. Maybe we could figure that out in class. The result then is different with the fifht article having the longest subtitle.